Renamed the controller named 'users' to the usersController
Updated imports for that controller
